+++ sourceCode/utils/sha256.cpp
@@ -158,7 +158,7 @@ std::string sha256(std::string input)
  
     SHA256 ctx = SHA256();
     ctx.init();
-    ctx.update( (unsigned char*)input.c_str(), input.length());
+    ctx.update( (unsigned char*)input.c_str(), (unsigned int)input.length());
     ctx.final(digest);
  
     char buf[2*SHA256::DIGEST_SIZE+1];
